fixed initialisation of action controls

+++ XRNeckSaferApp/ActionControls/Designer/ActionPropertyGroup.cs
@@ -3,7 +3,7 @@
     public class ActionPropertyGroup
     {
         public string Name { get; set; }
-        public int Order { get; set; }
+        public short Order { get; set; }
 
         public override string ToString()
         {
